move dataflatten
