+++ ee/dataflatten/flatten.go
