+++ ee/dataflatten/json.go
