+++ ee/dataflatten/testdata/animals.json
