+++ ee/dataflatten/testdata/animals.jsonl
