+++ ee/dataflatten/testdata/animals.plist
