+++ ee/dataflatten/testdata/animals.xml
