+++ ee/dataflatten/testdata/nested/inner.json
