+++ pkg/osquery/table/launcher_db_info.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 
 	"github.com/kolide/launcher/pkg/agent"
-	"github.com/kolide/launcher/pkg/dataflatten"
+	"github.com/kolide/launcher/ee/dataflatten"
 	"github.com/kolide/launcher/pkg/osquery/tables/dataflattentable"
 	"github.com/kolide/launcher/pkg/osquery/tables/tablehelpers"
 	"github.com/osquery/osquery-go/plugin/table"
